feat: 实现终端UI界面及核心功能
新增Textual终端界面，包含以下主要组件：
- ScoreTable：展示指数打分表格，支持颜色区分估值水平
- FactorDetailScreen：模态弹窗展示因子详情和LLM解读
- ActionBar：底部快捷键提示栏
- IndexScoreApp：主应用框架，集成数据加载、刷新、报告生成功能

新增数据加载流程：
- 异步拉取指数数据
- 计算价格位置和综合打分
- 调用LLM生成解读文本
- 通过消息机制更新UI

新增快捷键支持：
- R键刷新数据
- Enter键查看详情
- G键生成报告
- Q键退出应用

同时更新相关测试用例和文档说明

+++ src/index_score/config/loader.py
@@ -121,7 +121,7 @@ def _parse_score_ranges(items: list[dict[str, Any]]) -> list[ScoreRange]:
     return [
         ScoreRange(
             max_percentile=float(item["max_percentile"]),
-            score=int(item["score"]),
+            score=float(item["score"]),
         )
         for item in items
     ]
